Move yieldForEachJob to legacy and deprecate
